feat: adicionar restirção de rotas

+++ projeto_laravel/routes/api.php
@@ -19,15 +19,15 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::middleware('auth:sanctum')->group(function () {
-    Route::get('/produto/{id}', [ProdutoController::class, 'show']);
-    Route::post('/produto', [ProdutoController::class, 'store']);
-    Route::put('/produto/{id}', [ProdutoController::class, 'update']);
-    Route::delete('/produto/{id}', [ProdutoController::class, 'remove']);
+    Route::get('/produto/{id}', [ProdutoController::class, 'show'])->middleware('ability:client,admin');
+    Route::post('/produto', [ProdutoController::class, 'store'])->middleware('ability:client,admin');
+    Route::put('/produto/{id}', [ProdutoController::class, 'update'])->middleware('ability:client,admin');
+    Route::delete('/produto/{id}', [ProdutoController::class, 'remove'])->middleware('ability:admin');
 
-    Route::apiResource('forums', ForumController::class);
+    Route::apiResource('forums', ForumController::class)->middleware('ability:manager,admin');
 
-    Route::apiResource('usuarios', UsuarioController::class);
-    Route::get('usuarios/{usuario}/produtos', [UsuarioController::class, 'getProdutos']);
+    Route::apiResource('usuarios', UsuarioController::class)->middleware('ability:client,admin');
+    Route::get('usuarios/{usuario}/produtos', [UsuarioController::class, 'getProdutos'])->middleware('ability:client,admin');
 });
 
 Route::post('/login', [LoginController::class, 'login'])->name('login');
